requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
spacy
spacy-transformers
pytesseract
opencv-python
pandas
python-docx
pdfplumber
faker
scikit-learn
streamlit
python-multipart
python-dotenv
typer
requests
matplotlib
seaborn
numpy
pillow
